requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
konlpy
openai
python-dotenv